Moved zip folder service to its own file, managed network error from uploadManager and changed how handle errors from upload manager
