+++ src/app/core/services/stream.service.ts
@@ -1,7 +1,3 @@
-import createZipReadable from 'app/drive/services/download.service/downloadFolder/zipStream';
-// import streamSaver from 'streamsaver';
-import { loadWritableStreamPonyfill } from 'app/network/download';
-
 type BinaryStream = ReadableStream<Uint8Array>;
 
 export async function binaryStreamToBlob(stream: BinaryStream, mimeType?: string): Promise<Blob> {
@@ -133,93 +129,4 @@ export function streamFileIntoChunks(
       buffer = new Uint8Array(buffer.slice(chunkSize));
     },
   });
-}
-
-type FlatFolderZipOpts = {
-  abortController?: AbortController;
-  progress?: (loadedBytes: number) => void;
-};
-
-type AddFileToZipFunction = (name: string, source: ReadableStream<Uint8Array>) => void;
-type AddFolderToZipFunction = (name: string) => void;
-
-interface ZipStream {
-  addFile: AddFileToZipFunction;
-  addFolder: AddFolderToZipFunction;
-  stream: ReadableStream<Uint8Array>;
-  end: () => void;
-}
-
-export class FlatFolderZip {
-  private finished!: Promise<void>;
-  private zip: ZipStream;
-  private abortController?: AbortController;
-
-  constructor(folderName: string, opts: FlatFolderZipOpts) {
-    this.zip = createFolderWithFilesWritable();
-    this.abortController = opts.abortController;
-
-    const passThrough = opts.progress ? buildProgressStream(this.zip.stream, opts.progress) : this.zip.stream;
-
-    const isFirefox = navigator.userAgent.indexOf('Firefox') != -1;
-
-    if (isFirefox) {
-      loadWritableStreamPonyfill().then(() => {
-        // streamSaver.WritableStream = window.WritableStream;
-        // this.finished = passThrough.pipeTo(streamSaver.createWriteStream(folderName + '.zip'), {
-        //   signal: opts.abortController?.signal,
-        // });
-      });
-    } else {
-      // this.finished = passThrough.pipeTo(streamSaver.createWriteStream(folderName + '.zip'), {
-      //   signal: opts.abortController?.signal,
-      // });
-    }
-  }
-
-  addFile(name: string, source: ReadableStream<Uint8Array>): void {
-    if (this.abortController?.signal.aborted) return;
-
-    this.zip.addFile(name, source);
-  }
-
-  addFolder(name: string): void {
-    if (this.abortController?.signal.aborted) return;
-
-    this.zip.addFolder(name);
-  }
-
-  async close(): Promise<void> {
-    if (this.abortController?.signal.aborted) return;
-
-    this.zip.end();
-    await this.finished;
-  }
-
-  abort(): void {
-    this.abortController?.abort();
-  }
-}
-
-function createFolderWithFilesWritable(): ZipStream {
-  let controller;
-
-  const zipStream = createZipReadable({
-    start(ctrl) {
-      controller = ctrl;
-    },
-  });
-
-  return {
-    addFile: (name: string, source: ReadableStream<Uint8Array>): void => {
-      controller.enqueue({ name, stream: () => source });
-    },
-    addFolder: (name: string): void => {
-      controller.enqueue({ name, directory: true });
-    },
-    stream: zipStream,
-    end: () => {
-      controller.close();
-    },
-  };
 }
